+++ CHANGES.md
@@ -2,6 +2,7 @@
 # Unreleased
 
 * add `interpolate=True/False` to `.point()` methods to allow interpolation of surrounding pixels
+* add `pixel_location` property to `PointData` model
 
 # 7.5.1 (2025-03-19)