+++ CHANGES.md
@@ -1,6 +1,9 @@
 
 # Unreleased
 
+* add `interpolate=True/False` to `.point()` methods to allow interpolation of surrounding pixels
+* add `pixel_location` property to `PointData` model
+
 # 7.5.1 (2025-03-19)
 
 * fix `utils.get_array_statistics` method to avoid `ZeroDivisionError` when there is no valid pixel