+++ CHANGES.md
@@ -43,6 +43,8 @@
         >> float32
     ```
 
+* update pystac dependency to `>=1.9,<2.0`
+
 # 7.5.1 (2025-03-19)
 
 * fix `utils.get_array_statistics` method to avoid `ZeroDivisionError` when there is no valid pixel